Move SuccessCriteriaEvaluator to orchestration/ layer (GH #6832)
- git mv enhanced_orchestration/success_criteria.py → orchestration/success_criteria.py
- Replace original with backward-compat re-export shim (5 public symbols)
- Fix enhanced_orchestration/types.py TYPE_CHECKING import to new path
- Wire SuccessCriteriaEvaluator into WorkflowExecutor as optional kwarg;
  async evaluation runs in execute_coordinated_workflow after status is set
- Export SuccessCriteriaEvaluator from orchestration/__init__.py
- Add GH #6832 design note to CollaborationCoordinator docstring
