+++ autobot-backend/enhanced_orchestration/collaboration_coordinator.py
@@ -21,6 +21,11 @@ class CollaborationCoordinator:
 
     Accept a redis_factory for testability (#6401); defaults to the production
     async client so production callers need no changes.
+
+    Design note (GH #6832): although this module has a single caller (WorkflowRunner),
+    the redis_factory injection makes it worth keeping separate — tests can pass a mock
+    factory without patching module-level globals. Inline only if a second caller never
+    materialises.
     """
 
     def __init__(self, redis_factory: Callable = get_async_redis_client) -> None:
